+++ src/kaocha/plugin/junit_xml.clj
@@ -1,19 +1,26 @@
 (ns kaocha.plugin.junit-xml
-  (:require [kaocha.plugin :as plugin :refer [defplugin]]
+  (:require [clojure.java.io :as io]
             [kaocha.core-ext :refer :all]
-            [kaocha.testable :as testable]
-            [kaocha.history :as history]
-            [kaocha.result :as result]
-            [kaocha.plugin.junit-xml.xml :as xml]
-            [clojure.java.io :as io]
-            [kaocha.report :as report]
             [kaocha.hierarchy :as hierarchy]
-            [kaocha.output :as output]))
+            [kaocha.plugin :as plugin :refer [defplugin]]
+            [kaocha.plugin.junit-xml.xml :as xml]
+            [kaocha.report :as report]
+            [kaocha.result :as result]
+            [kaocha.testable :as testable]
+            [clojure.string :as str])
+  (:import java.time.Instant))
 
 (defn inst->iso8601 [inst]
   (.. java.time.format.DateTimeFormatter/ISO_LOCAL_DATE_TIME
       (withZone (java.time.ZoneId/systemDefault))
       (format (.truncatedTo inst java.time.temporal.ChronoUnit/SECONDS))))
+
+;; The ESC [ is followed by any number (including none) of "parameter bytes" in
+;; the range 0x30–0x3F (ASCII 0–9:;<=>?), then by any number of "intermediate
+;; bytes" in the range 0x20–0x2F (ASCII space and !"#$%&'()*+,-./), then finally
+;; by a single "final byte" in the range 0x40–0x7E (ASCII @A–Z[\]^_`a–z{|}~).
+(defn strip-ansi-sequences [s]
+  (str/replace s #"\x1b\[[\x30-x3F]*[\x20-\x2F]*[\x40-\x7E]" ""))
 
 (defn test-seq [testable]
   (cons testable (mapcat test-seq (::result/tests testable))))
@@ -23,19 +30,18 @@
           (test-seq testable)))
 
 (defn time-stat [testable]
-  (let [duration (:kaocha.plugin.profiling/duration testable)]
+  (let [duration (:kaocha.plugin.profiling/duration testable 0)]
     (when duration
       {:time (format "%.6f" (/ duration 1e9))})))
 
 (defn stats [testable]
   (let [tests      (test-seq testable)
         totals     (result/totals (::result/tests testable))
-        start-time (:kaocha.plugin.profiling/start testable)]
-    (cond-> {:skipped (count (filter ::testable/skip tests))
-             :errors   (::result/error totals)
-             :failures (::result/fail totals)
-             :tests    (::result/count totals)}
-      start-time (assoc :timestamp (inst->iso8601 start-time)))))
+        start-time (:kaocha.plugin.profiling/start testable (Instant/now))]
+    {:errors   (::result/error totals)
+     :failures (::result/fail totals)
+     :tests    (::result/count totals)
+     :timestamp (inst->iso8601 start-time)}))
 
 (defn test-name [test]
   (let [id (::testable/id test)]
@@ -44,8 +50,9 @@
          (name id))))
 
 (defn failure-message [events]
-  (with-out-str
-    (run! report/print-expr events)))
+  (strip-ansi-sequences
+   (with-out-str
+     (run! report/print-expr events))))
 
 (defn error-type [events]
   (let [e (first (filter #(throwable? (:actual %)) events))]
@@ -57,12 +64,10 @@
          :or             {pass 0 fail 0 error 0}} test]
     {:tag     :testcase
      :attrs   (merge {:name       (test-name test)
-                      :classname  (namespace id)
-                      :assertions (+ pass fail error)}
+                      :classname  (namespace id)}
                      (time-stat test))
      :content (keep identity
                     [(cond
-                       skip        {:tag :skipped}
                        (> error 0) (let [events (filter (comp #{:error} :type) events)
                                          message (some-> events first :actual (.getMessage))
                                          trace (failure-message events)]
@@ -75,11 +80,8 @@
                                                           (filter #(and (hierarchy/fail-type? %)
                                                                         (not (= :error (:type %)))))
                                                           failure-message)
-                                            :type    (first (remove #{:default} (map report/assertion-type events)))}})
-                     (when-let [out (:kaocha.plugin.capture-output/output test)]
-                       (when (seq out)
-                         {:tag     :system-out
-                          :content [out]}))])}))
+                                            :type    (str (first (remove #{:default} (map report/assertion-type events))))}})
+                     ])}))
 
 (defn suite->xml [suite index]
   (let [id (::testable/id suite)]
@@ -88,14 +90,24 @@
                  :id index
                  :hostname "localhost"}
                 (merge (stats suite) (time-stat suite))
-                (cond-> (qualified-ident? id) (assoc :package (namespace id))))
-     :content (map testcase->xml (leaf-tests suite))}))
-
+                (assoc :package (if (qualified-ident? id)
+                                  (namespace id)
+                                  "")))
+     :content (concat
+               [{:tag :properties}]
+               (map testcase->xml (leaf-tests suite))
+               [{:tag :system-out
+                 :content (->> suite
+                               test-seq
+                               (keep :kaocha.plugin.capture-output/output)
+                               (remove #{""})
+                               (map strip-ansi-sequences))}
+                {:tag :system-err}])}))
 
 (defn result->xml [result]
   (let [suites (::result/tests result)]
-    {:tag     :testuites
-     :attrs   (stats result)
+    {:tag     :testsuites
+     :attrs   {}
      :content (map suite->xml suites (range))}))
 
 (defn write-junit-xml [filename result]