+++ test/unit/kaocha/plugin/junit_xml/xml_test.clj
@@ -1,7 +1,8 @@
 (ns kaocha.plugin.junit-xml.xml-test
   (:require [clojure.test :refer :all]
             [clojure.xml :as cxml]
-            [kaocha.plugin.junit-xml.xml :as xml]))
+            [kaocha.plugin.junit-xml.xml :as xml]
+            [clojure.java.io :as io]))
 
 (deftest emit-test
   (testing "escapes text"
@@ -13,3 +14,12 @@
     (is (= "<?xml version='1.0' encoding='UTF-8'?>\n<foo hello=\"'twas brillig &amp; the &quot;slithy&quot; toves\"/>\n"
            (xml/emit-str {:tag :foo
                           :attrs {:hello "'twas brillig & the \"slithy\" toves"}})))))
+
+(deftest validate-test
+  (is (= (xml/validate "<foo></foo>" (io/resource "kaocha/junit_xml/JUnit.xsd"))
+         {:valid? false,
+          :line-number 1,
+          :column-number 6,
+          :public-id nil,
+          :system-id nil,
+          :message "cvc-elt.1: Cannot find the declaration of element 'foo'."})))